+++ tpv/core/helpers.py
@@ -13,7 +13,9 @@ GIGABYTES = 1024.0**3
 
 
 def get_dataset_size(dataset):
-    return float(dataset.file_size)
+    # calculate_size would mark file_size column as dirty
+    # and may have unintended consequences
+    return float(dataset.get_size(calculate_size=False))
 
 
 def sum_total(prev, current):
@@ -112,3 +114,14 @@ def tool_version_gte(tool, version):
 
 def tool_version_gt(tool, version):
     return parse_version(tool.version) > parse_version(version)
+
+
+def get_dataset_attributes(datasets):
+    # Return a dictionary of dataset ids and their object store ids
+    # and file sizes in bytes for all input datasets in a job
+    return {
+        i.dataset.dataset.id: {
+            'object_store_id': i.dataset.dataset.object_store_id,
+            'size': get_dataset_size(i.dataset.dataset)}
+        for i in datasets or {}
+    }